Fix out-of-memory in the 8,000-image run
The full STL-10 run died when dense-SIFT extraction started: DINOv2
features were still resident and extraction held two float32 copies.
Measured at full SIFT scale (8,000 x 729 x 128), peak RSS is now 5.0 GB.

- extractors keep descriptors in float16 and stack into one preallocated array
- sweep frees each extractor's features (and encodings) before the next
- VLAD intra-norm and power/L2 normalisation run in place, in row blocks
  (VLAD K=256 encode peak 4.6 GB -> 2.9 GB)
- cache.save avoids a redundant float16 copy
- record GPU name even before torch is first imported

+++ bovw/extract/foundation.py
@@ -73,5 +73,6 @@ class FoundationExtractor:
                 h = out.last_hidden_state if self.layer == -1 else out.hidden_states[self.layer]
                 h = h.float().cpu().numpy()
                 cls.append(h[:, 0])
-                patches.extend(h[:, self.spec.n_prefix:])  # (P, D) per image
-        return LocalFeatures.from_list(patches, global_=np.concatenate(cls).astype(np.float32))
+                patches.extend(h[:, self.spec.n_prefix:].astype(np.float16))  # (P, D) per image
+        return LocalFeatures.from_list(patches, global_=np.concatenate(cls).astype(np.float32),
+                                       dtype=np.float16)
